+++ debugoverlay/src/main/java/com/ms_square/debugoverlay/modules/CpuUsageViewModule.java
@@ -17,7 +17,7 @@ import java.util.Locale;
 
 public class CpuUsageViewModule extends BaseViewModule<CpuUsage> {
 
-    private static final String TAG = CpuUsageViewModule.class.getSimpleName();
+    private static final String TAG = "CpuUsageViewModule";
 
     private static final DecimalFormat DECIMAL_FORMAT = new DecimalFormat("0.0",
             DecimalFormatSymbols.getInstance(Locale.ENGLISH));
@@ -44,7 +44,7 @@ public class CpuUsageViewModule extends BaseViewModule<CpuUsage> {
             }
 
             if (cpuUsageTextView != null) {
-                StringBuilder builder = new StringBuilder("cpu: ");
+                StringBuilder builder = new StringBuilder("cpu(%): ");
                 cpuUsageTextView.setText(builder.append(totalCpuUsage).append(" ").append(myPidCpuUsage).toString());
             }
         }
@@ -53,7 +53,7 @@ public class CpuUsageViewModule extends BaseViewModule<CpuUsage> {
     @Override
     public View createView(ViewGroup root, @ColorInt int textColor, float textSize, float textAlpha) {
         View view = LayoutInflater.from(root.getContext()).inflate(layoutResId, root, false);
-        cpuUsageTextView = (TextView) view.findViewById(R.id.debugoverlay_overlay_text);
+        cpuUsageTextView = view.findViewById(R.id.debugoverlay_overlay_text);
         cpuUsageTextView.setTextColor(textColor);
         cpuUsageTextView.setTextSize(textSize);
         cpuUsageTextView.setAlpha(textAlpha);